requirements.txt
# FastAPI and web server
fastapi==0.108.0
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Gaia API and HTTP client
httpx==0.25.2
asyncio==3.4.3

# Web3 and blockchain
web3==6.11.3

# Utilities
tenacity==8.2.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development tools
black==23.3.0
isort==5.12.0
mypy==1.3.0
flake8==6.0.0